MULTIPOLYGON, auth token, sf, LoadCountries
Converted POLYGON geometry to MULTIPOLYGON

added sf as a dependency in DESCRIPTION so that the package is attached (does not happen in import)

added missing file documentation and updated README to accommodate oauth tokens for Philippines and made changes to both LoadPhilippines and LoadCountries to work with this. Added interactiveMode option to LoadCountries which will remove LoadPhilippines from the list of sources to load data from if not in interactive model and no googledrive authorization token available.